Some cleanup and reorganization of avi code AVIodgAuto maintains ODG information automatically AVIodgReadonly and AVIodgImplicit are handwritten two-phase version, used just for comparisons.
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@11114 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ apps/avi/libMat/Material.h
@@ -69,7 +69,7 @@ public:
   static const double PERT;
   static const double DET_MIN;
 
-  static const int MAT_SIZE = 9;
+  static const size_t MAT_SIZE = 9;
   static const size_t NDF;
   static const size_t NDM;
